refactor(hooks/builtins): split each builtin into its own file
builtins.go now only holds the package doc, name constants, Register,

AgentDefaults / ApplyAgentDefaults, and the shared turnStartContext helper.

Each builtin (add_date, add_environment_info, add_prompt_files) lives in

its own file, making it easier to read, locate, and add new builtins.

Public API and behaviour are unchanged; tests pass as-is.
