+++ pkg/hooks/builtins/builtins.go
@@ -10,17 +10,15 @@
 // AddDate and AddPromptFiles target turn_start so they recompute every
 // turn. AddEnvironmentInfo targets session_start because cwd / OS / arch
 // don't change during a session.
+//
+// Each builtin lives in its own file (add_date.go, add_environment_info.go,
+// add_prompt_files.go); this file holds the shared registration plumbing.
 package builtins
 
 import (
-	"context"
 	"fmt"
-	"log/slog"
-	"strings"
-	"time"
 
 	"github.com/docker/docker-agent/pkg/hooks"
-	"github.com/docker/docker-agent/pkg/session"
 )
 
 // Builtin hook names. They match the `command` field of a
@@ -97,42 +95,8 @@ func ApplyAgentDefaults(cfg *hooks.Config, d AgentDefaults) *hooks.Config {
 }
 
 // turnStartContext wraps additional context as a turn_start output.
+// Shared by builtins that contribute per-turn context (add_date,
+// add_prompt_files).
 func turnStartContext(content string) *hooks.Output {
 	return hooks.NewAdditionalContextOutput(hooks.EventTurnStart, content)
-}
-
-// addDate emits today's date as turn_start additional context.
-func addDate(_ context.Context, _ *hooks.Input, _ []string) (*hooks.Output, error) {
-	return turnStartContext("Today's date: " + time.Now().Format("2006-01-02")), nil
-}
-
-// addEnvironmentInfo emits cwd / git / OS / arch info as session_start
-// additional context. No-op when Cwd is empty.
-func addEnvironmentInfo(_ context.Context, in *hooks.Input, _ []string) (*hooks.Output, error) {
-	if in == nil || in.Cwd == "" {
-		return nil, nil
-	}
-	return hooks.NewAdditionalContextOutput(hooks.EventSessionStart, session.GetEnvironmentInfo(in.Cwd)), nil
-}
-
-// addPromptFiles reads each filename in args (relative to Input.Cwd) and
-// joins their contents into turn_start additional context. Missing files
-// are logged and skipped; surviving files still contribute.
-func addPromptFiles(_ context.Context, in *hooks.Input, args []string) (*hooks.Output, error) {
-	if in == nil || in.Cwd == "" || len(args) == 0 {
-		return nil, nil
-	}
-	var parts []string
-	for _, name := range args {
-		additional, err := session.ReadPromptFiles(in.Cwd, name)
-		if err != nil {
-			slog.Warn("reading prompt file", "file", name, "error", err)
-			continue
-		}
-		parts = append(parts, additional...)
-	}
-	if len(parts) == 0 {
-		return nil, nil
-	}
-	return turnStartContext(strings.Join(parts, "\n\n")), nil
 }
